+++ cmd/provisioner/commands/root.go
@@ -9,9 +9,11 @@ import (
 	"k8s-pv-provisioner/cmd/provisioner/controllers/pvc"
 	"os"
 	"strconv"
+	"strings"
 
 	"github.com/spf13/cobra"
 	v1 "k8s.io/api/core/v1"
+	storage_v1 "k8s.io/api/storage/v1"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/client-go/kubernetes"
@@ -28,14 +30,14 @@ var (
 		Short: "Custom persistent volume provisioner for Kubernetes",
 	}
 
-	//storageClassName is a name of the k8s storage-class for what the provisioner will work
-	storageClassName string
-	/*storageAssetRoot is the directory on file system under which a new storage assets will be created for provisioned PV
-	or will be looked and deleted if the PV should be deprovisioned*/
+	/*storageClassName is a comma separated names of k8s storage-classes for which the provisioner should work*/
+	storageClassNames string
+	/*storageAssetRoot is the directory on file system under the which a new storage assets will be created or deleted
+	by provisioner*/
 	storageAssetRoot string
-	//kubectlConfig is path where actual config file for kubectl is placed for the k8s
+	/*kubectlConfig is path where actual config file for kubectl is placed for the k8s*/
 	kubectlConfig string
-	//verbosityLogging is logging level for app
+	/*verbosityLogging is logging level for app*/
 	verbosityLogging int
 )
 
@@ -45,15 +47,15 @@ func init() {
 		Short: "starts the watching process for provision/deprovision persistentVolume of K8s",
 	}
 
-	serveCmd.Flags().StringVar(&storageClassName, "storage-class", "", "name of storage class to watch for (requred)")
+	serveCmd.Flags().StringVar(&storageClassNames, "storage-classes", "", "comma separated list of storage class names to watch for (requred)")
 	serveCmd.Flags().StringVar(&storageAssetRoot, "storage-asset-root", "", "directory where assets will be created  (requred)")
-	serveCmd.MarkFlagRequired("storage-class")
+	serveCmd.MarkFlagRequired("storage-classes")
 	serveCmd.MarkFlagRequired("storage-asset-root")
 	serveCmd.Run = run
 
 	rootCmd.AddCommand(serveCmd)
 	rootCmd.PersistentFlags().StringVarP(&kubectlConfig, "kubectl-config", "c", "", "path to kubectl's config")
-	rootCmd.PersistentFlags().IntVar(&verbosityLogging, "v", 0, "logging verbosity")
+	rootCmd.PersistentFlags().IntVar(&verbosityLogging, "v", 0, "logging verbosity (0..2)")
 	rootCmd.PersistentPreRun = persistentPreRun
 }
 
@@ -65,14 +67,45 @@ func Execute() {
 	}
 }
 
+//Initializing the klog
 func persistentPreRun(cmd *cobra.Command, args []string) {
-	//Initing the klog
 	flags := flag.NewFlagSet("Logging flags", flag.ExitOnError)
 	klog.InitFlags(flags)
 
 	flags.Set("logtostderr", "true")
 	flags.Set("v", strconv.Itoa(verbosityLogging))
 	flags.Parse(nil)
+}
+
+func selectClasses(source []storage_v1.StorageClass, patterns []string) ([]storage_v1.StorageClass, error) {
+	/* Parsing storage class names gathered from the CLI-arg to keys of the map which will
+	track what storage classes will be found in the cluster*/
+	wanted := make(map[string]bool)
+	for _, item := range patterns {
+		if len(item) > 0 {
+			wanted[item] = false
+		}
+	}
+	result := make([]storage_v1.StorageClass, 0)
+	//Parsing storage class list gathered from the cluster
+	for _, storageClass := range source {
+		if _, present := wanted[storageClass.Name]; present {
+			wanted[storageClass.Name] = true
+			result = append(result, storageClass)
+		}
+	}
+	//Checkng the result of the operation
+	lost := make([]string, 0)
+	for key, value := range wanted {
+		if !value {
+			lost = append(lost, key)
+		}
+	}
+	if len(lost) > 0 {
+		return nil, fmt.Errorf("Could not find specified storage classes: %v", strings.Join(lost, ", "))
+	}
+
+	return result, nil
 }
 
 func run(cmd *cobra.Command, args []string) {
@@ -87,7 +120,6 @@ func run(cmd *cobra.Command, args []string) {
 		klog.Info("Trying to use config specifyied as file path")
 		config, err = clientcmd.BuildConfigFromFlags("", kubectlConfig)
 	}
-
 	if err != nil {
 		klog.Fatal(err.Error())
 	}
@@ -102,22 +134,20 @@ func run(cmd *cobra.Command, args []string) {
 	appConfig.StorageAssetRoot = storageAssetRoot
 	appConfig.Clientset = clientset
 
-	storageClasses, err := clientset.StorageV1().StorageClasses().List(meta_v1.ListOptions{})
+	clusterStorageClasses, err := clientset.StorageV1().StorageClasses().List(meta_v1.ListOptions{})
 	if err != nil {
 		klog.Fatal("Could not fetch list of storage classes")
 	}
 
-	for _, item := range storageClasses.Items {
-		if item.Name == storageClassName {
-			appConfig.ParseStorageClass(&item)
-			break
-		}
+	selected, err := selectClasses(clusterStorageClasses.Items, strings.Split(storageClassNames, ","))
+	if err != nil {
+		klog.Fatal(err)
 	}
-	if appConfig.StorageClass.Name == "" {
-		klog.Fatalf("Could not find specified storage class: %v", storageClassName)
+	for _, storageClass := range selected {
+		appConfig.ParseStorageClass(&storageClass)
 	}
 
-	//Preparing stuff for PVC controller
+	//Preparation steps for PVC controller
 	pvcListWatcher := cache.NewListWatchFromClient(clientset.CoreV1().RESTClient(), "persistentvolumeclaims", meta_v1.NamespaceAll, fields.Everything())
 	pvcQueue := workqueue.NewRateLimitingQueue(workqueue.DefaultControllerRateLimiter())
 	pvcIndexer, pvcInformer := cache.NewIndexerInformer(pvcListWatcher, &v1.PersistentVolumeClaim{}, 0, cache.ResourceEventHandlerFuncs{
@@ -130,8 +160,10 @@ func run(cmd *cobra.Command, args []string) {
 			}
 		},
 	}, cache.Indexers{})
+	pvcCtrl := controllers.NewController("PersistentVolumeClaim", pvcQueue, pvcIndexer, pvcInformer)
+	pvcCtrl.ItemHandler = pvc.Handler
 
-	//Preparing stuff for PV controller
+	//Preparation steps for PV controller
 	pvListWatcher := cache.NewListWatchFromClient(clientset.CoreV1().RESTClient(), "persistentvolumes", meta_v1.NamespaceNone, fields.Everything())
 	pvQueue := workqueue.NewRateLimitingQueue(workqueue.DefaultControllerRateLimiter())
 	pvIndexer, pvInformer := cache.NewIndexerInformer(pvListWatcher, &v1.PersistentVolume{}, 0, cache.ResourceEventHandlerFuncs{
@@ -144,17 +176,12 @@ func run(cmd *cobra.Command, args []string) {
 			}
 		},
 	}, cache.Indexers{})
-
-	var pvcCtrl = controllers.NewController("PersistentVolumeClaim", pvcQueue, pvcIndexer, pvcInformer)
-	pvcCtrl.ItemHandler = pvc.Handler
-
-	var pvCtrl = controllers.NewController("PersistentVolume", pvQueue, pvIndexer, pvInformer)
+	pvCtrl := controllers.NewController("PersistentVolume", pvQueue, pvIndexer, pvInformer)
 	pvCtrl.ItemHandler = pv.Handler
 
 	//Starting the controllers with one stop-channel
 	stop := make(chan struct{})
 	defer close(stop)
-
 	go pvcCtrl.Run(stop)
 	go pvCtrl.Run(stop)