+++ cmd/provisioner/controllers/pvc/pvc_checkers.go
@@ -3,12 +3,14 @@ package pvc
 import (
 	"fmt"
 	"k8s-pv-provisioner/cmd/provisioner/config"
+	"strings"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/klog"
 )
 
+//TODO: Refactor this module at all
 var predicates = []func(*v1.PersistentVolumeClaim) bool{
 	HasProperStorageClassName,
 	HasProperProvisionerAnnotation,
@@ -17,7 +19,7 @@ var predicates = []func(*v1.PersistentVolumeClaim) bool{
 
 func IsBoundPVC(pvc *v1.PersistentVolumeClaim) bool {
 	if pvc.Spec.VolumeName != "" {
-		klog.V(2).Infof("PersistentVolumeClaim: %v has been bound to a volume according to Spec.VolumeName", pvc.Name)
+		klog.V(2).Infof("PersistentVolumeClaim: %v already had been bound to the volume according to Spec.VolumeName", pvc.Name)
 		return true
 	}
 
@@ -25,11 +27,16 @@ func IsBoundPVC(pvc *v1.PersistentVolumeClaim) bool {
 }
 
 func HasProperStorageClassName(pvc *v1.PersistentVolumeClaim) bool {
-	if *pvc.Spec.StorageClassName == appConfig.StorageClass.Name {
+	if _, present := appConfig.StorageClasses[*pvc.Spec.StorageClassName]; present {
 		return true
 	}
 
-	klog.V(2).Infof("PersistentVolumeClaim: %v should be provisioned by another storageClass than: %v", pvc.Name, appConfig.StorageClass.Name)
+	storageClasses := make([]string, 0)
+	for key := range appConfig.StorageClasses {
+		storageClasses = append(storageClasses, key)
+	}
+
+	klog.V(2).Infof("PersistentVolumeClaim: %v should be provisioned by another storageClass rather than on of: %v", pvc.Name, strings.Join(storageClasses, ", "))
 	return false
 }
 
@@ -44,11 +51,12 @@ func IsSelectorsListEmpty(pvc *v1.PersistentVolumeClaim) bool {
 
 func HasProperProvisionerAnnotation(pvc *v1.PersistentVolumeClaim) bool {
 	value, ok := pvc.Annotations[config.AnnotationStorageProvisioner]
-	if ok && value == appConfig.StorageClass.Provisioner {
+	sc := appConfig.StorageClasses[*pvc.Spec.StorageClassName]
+	if ok && value == sc.Provisioner {
 		return true
 	}
 
-	klog.V(2).Infof("PersistentVolumeClaim: %v does not have annotation specifying onto the provisioner: %v", pvc.Name, config.AnnotationStorageProvisioner)
+	klog.V(2).Infof("PersistentVolumeClaim: %v does not have needed annotation: %v", pvc.Name, config.AnnotationStorageProvisioner)
 	return false
 }