+++ cmd/provisioner/storage/storage_utils.go
@@ -3,7 +3,6 @@ package storage
 import (
 	"fmt"
 	"os"
-	"path"
 	"strconv"
 	"strings"
 
@@ -32,26 +31,28 @@ func ChooseAssetOwner(pvc *v1.PersistentVolumeClaim) (int, int) {
 	var uid, gid int
 	var err error
 
-	tmp, ok := pvc.Annotations[config.AnnotationOwnerNewAssetUID]
+	currentStorageClass := appConfig.StorageClasses[*pvc.Spec.StorageClassName]
+
+	value, ok := pvc.Annotations[config.AnnotationOwnerNewAssetUID]
 	if ok {
-		uid, err = CastToInt(tmp)
+		uid, err = CastToInt(value)
 		if err != nil {
-			uid = appConfig.StorageClass.DefaultOwnerAssetUID
-			klog.Warningf("PersistentVolumeClaim: %v annotation: %v could not parse value: %v: %v", pvc.Name, config.AnnotationOwnerNewAssetUID, tmp, err)
+			uid = currentStorageClass.DefaultOwnerAssetUID
+			klog.Warningf("PersistentVolumeClaim: %v annotation: %v could not parse value: %v: %v", pvc.Name, config.AnnotationOwnerNewAssetUID, value, err)
 		}
 	} else {
-		uid = appConfig.StorageClass.DefaultOwnerAssetUID
+		uid = currentStorageClass.DefaultOwnerAssetUID
 	}
 
-	tmp, ok = pvc.Annotations[config.AnnotationOwnerNewAssetGID]
+	value, ok = pvc.Annotations[config.AnnotationOwnerNewAssetGID]
 	if ok {
-		gid, err = CastToInt(tmp)
+		gid, err = CastToInt(value)
 		if err != nil {
-			gid = appConfig.StorageClass.DefaultOwnerAssetGID
-			klog.Warningf("PersistentVolumeClaim: %v annotation: %v could not parse value: %v: %v", pvc.Name, config.AnnotationOwnerNewAssetGID, tmp, err)
+			gid = currentStorageClass.DefaultOwnerAssetGID
+			klog.Warningf("PersistentVolumeClaim: %v annotation: %v could not parse value: %v: %v", pvc.Name, config.AnnotationOwnerNewAssetGID, value, err)
 		}
 	} else {
-		gid = appConfig.StorageClass.DefaultOwnerAssetGID
+		gid = currentStorageClass.DefaultOwnerAssetGID
 	}
 
 	return uid, gid
@@ -66,42 +67,36 @@ func CastToInt(value string) (int, error) {
 }
 
 //CreateStorageAsset is func which creates the storage asset
-func CreateStorageAsset(baseName string, uid, gid int) (string, error) {
+func CreateStorageAsset(assetPath string, uid, gid int) error {
 
-	appStorageAssetRoot := path.Join(appConfig.StorageAssetRoot, baseName)
-	classStorageAssetRoot := path.Join(appConfig.StorageClass.StorageAssetRoot, baseName)
-
-	if _, err := os.Stat(appStorageAssetRoot); err == nil {
-		return "", fmt.Errorf("Storage asset: %v already exists", appStorageAssetRoot)
+	if _, err := os.Stat(assetPath); err == nil {
+		return fmt.Errorf("Storage asset: %v already exists", assetPath)
 	}
 
-	if err := os.MkdirAll(appStorageAssetRoot, 0755); os.IsPermission(err) {
-		return "", err
+	if err := os.MkdirAll(assetPath, 0755); os.IsPermission(err) {
+		return err
 	}
 
-	klog.Infof("Storage asset: %v was successfully created", appStorageAssetRoot)
+	klog.Infof("Storage asset: %v was successfully created", assetPath)
 
-	if err := os.Chown(appStorageAssetRoot, uid, gid); err != nil {
-
-		return "", err
+	if err := os.Chown(assetPath, uid, gid); err != nil {
+		return err
 	}
 
-	klog.Infof("Storage asset: %v ownership was set as %v:%v", appStorageAssetRoot, uid, gid)
+	klog.Infof("Storage asset: %v ownership was set as %v:%v", assetPath, uid, gid)
 
-	return classStorageAssetRoot, nil
+	return nil
 }
 
 //DeleteStorageAsset is func deleting the storage asset
-func DeleteStorageAsset(pv *v1.PersistentVolume) error {
-	classStorageAsset := pv.Spec.PersistentVolumeSource.HostPath.Path
-	appStorageAsset := path.Join(appConfig.StorageAssetRoot, path.Base(classStorageAsset))
+func DeleteStorageAsset(assetPath string) error {
 
-	err := os.RemoveAll(appStorageAsset)
+	err := os.RemoveAll(assetPath)
 	if err != nil {
 		return err
 	}
 
-	klog.Infof("Storage asset: %v was successfully deleted", appStorageAsset)
+	klog.Infof("Storage asset: %v was successfully deleted", assetPath)
 
 	return nil
 }